html/semantics/embedded-content-0/the-canvas-element/2d.type.exists.html

<!DOCTYPE html>
<title>Canvas test: 2d.type.exists</title>
<meta name="author" content="Philip Taylor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../../../common/canvas-tests.js"></script>
<link rel="stylesheet" href="../../../../../common/canvas-tests.css">
<body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.type.html">type</a>.exists</h1>
<p class="desc">The 2D context interface is a property of 'window'</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.context.2d.type">context.2d.type</a>
</ul>
</div>
<p class="notes">Defined in "Web IDL" (draft)
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("The 2D context interface is a property of 'window'");
_addTest(function(canvas, ctx) {

_assert(window.CanvasRenderingContext2D, "window.CanvasRenderingContext2D");


});
</script>
